Move coffee optimisation related function to coffee_mode.py
